Convert Autocomplete to use mui-downshift
This is actually a simpler implementation. The styling works pretty well out of the box, and the wrapper wants to use objects instead of strings, with a toString converter. (I think downshift might have supported this, it just wasn't shown in the example from material-ui) The only thing I don't like about this at the moment is that the size changes when it disables/undisables, and when you select or deselect an item. This is because of the down arrow and x buttons getting added to the input.