Return the end addornments and force the column sizes
What I didn't like about the end addornments was that they were resizing the input whenever it disables, enables, or the user makes a selection, because of the addornments getting added and removed. I was able to fix this by forcing the columns to a particular size. Since the whole group is limited to a set size, this makes sense. I'm basically just dividing the overall width into two equal widths with a small space between.

For smaller screens, I'll just let it collapse to a two column layout. This makes it easier to use on a phone than it was.