Don't clear the textValue on props change
The new MuiDownshift autocomplete automatically clears the text when the autocomplete loses focus with no selection, so we don't need this anymore. This was causing the autocomplete to blank out when the user entered a docket number. When the user backspaces the selection, it should also emit a change event to null out the selection. Otherwise once they tab out the selection comes back.