Add redux for managing a global snackbar
Installed redux and react-redux and defined actions and reducers for the snackbar functionality. The snackbar will have open and close actions which toggle the visibility of the snackbar, as well as setting the message to be shown on the snackbar. I'm also going to include a type so we can show a different icon for different snack messages.