Open the success snackbar through the store
Implemented a separate snackbar component. This component is included directly in the app component. The component connects to the redux store and shows the bar with a message based on the snackbar state. The subscription component can then just dispatch an event to open the snackbar with the correct message and type.