Provide the complete dialog content
I didn't like having these split into separate content groups, as it made it hard to interact between the body and the footer. Also, the login dialog wrapped all of this in a form, which we might want for some dialogs but not necessarily all of them. This just means that dialog contents will have to include DialogTitle, DialogContent, and DialogActions components to define where each one is, but that shouldn't be a problem.