Open the login dialog via redux
Now that I have a dialog component that is connected to redux, the login dialog component just becomes a form component with dialog content tags to style the header, content, and footer correctly. The consuming context still needs to know when the user logs in successfully, so I'm passing an onSuccess handler into the action creator, which is then bound into the login dialog.